assignment completed

+++ app/Http/Controllers/User/AssignmentsController.php
@@ -33,13 +33,14 @@ class AssignmentsController extends Controller
         {
             $input = $request->all();
             $filepath = 'uploads/assignments';
-            $filename = $request->file('file')->getClientOriginalName();
+            $extension = $request->file('file')->getClientOriginalExtension();
+            $newFilename = str_slug($request->get('title'), "-").'.'.$extension;
             if(!(File::exists($filepath)))
             {
                 File::makeDirectory($filepath, 0775, true);
             }            
-            $request->file('file')->move($filepath, $filename);
-            $input['file'] = $filename;
+            $request->file('file')->move($filepath, $newFilename);
+            $input['file'] = $newFilename;
             Assignment::create($input);
             $response['data']['success'] = 'Saved Successfully';
             return $response;
@@ -52,7 +53,8 @@ class AssignmentsController extends Controller
         $subject = Hashids::connection('subject')->decode($request->get('subject'));
         $subject = Subject::find($subject)->first();
         $student = Student::where('user_id', Sentinel::getUser()->id)->first();
-        $assignment = $subject->assignments()->where('student_id', $student->id)->get();
+        $assignment = $subject->assignment()->where('student_id', $student->id)->get();
+        $response = [];
         foreach ($assignment as $key => $value) {
             $response[$key]['no'] = $key+1;
             $response[$key]['title'] = $value->title;
@@ -61,16 +63,30 @@ class AssignmentsController extends Controller
             $id = Hashids::connection('assignment')->encode($value->id);
             $response[$key]['action'] = 
                 '<div class="text-center"><form action="/assignment/'.$id.'" method="POST">'.csrf_field().'
-                    <button type="submit" class="btn btn-danger btn-xs btn-flat btn-delete">Delete</button>          
+                    <button type="submit" class="btn btn-danger btn-xs btn-flat btn-delete" id="deleteAssignment">Delete</button>          
                 </form></div>';
         }
         $data['data'] = $response;
         return response()->json($data, 200);
     }
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         $id = Hashids::connection('assignment')->decode($id);
-        Assignment::destroy($id);
-        
+        $assignment = Assignment::find($id)->first();
+        if($assignment->file != "")
+        {
+           if(File::exists('uploads/assignments/'.$assignment->file))
+            {
+            File::delete('uploads/assignments/'.$assignment->file);
+            } 
+        }
+        if($request->ajax())
+        {
+            
+             Assignment::destroy($id);
+             $response['success'] = 'Assignment Deleted Successfully';
+             return $response;   
+        }
+
     }
 }
